remove uint128 subdirectory from root and add uint128_t as a submodule to get the latest version as well as unit tests. Fix one build issue when using the uint128_t library, instead of the built-in type

+++ src/phase1.hpp
@@ -364,7 +364,7 @@ void* phase1_thread(void* arg)
 
                             // Rewrite left entry with just pos and offset, to reduce working space
                             Bits new_left_entry = Bits(
-                                (table_index == 1) ? entry->left_metadata : entry->read_posoffset,
+                                (table_index == 1) ? entry->left_metadata : uint128_t(entry->read_posoffset),
                                 (table_index == 1) ? k : pos_size + kOffsetSize);
 
                             new_left_entry.ToBytes(tmp_buf);
